requirements.txt
asgiref==3.8.1
Django
django-filter==24.3
djangorestframework==3.15.2
sqlparse==0.5.1
drf-yasg
